requirements.txt
aiogram
openai
python-dotenv
fastapi
uvicorn
jinja2
aiosqlite
